separate Fragment_Recruitment
