--- app/src/main/java/com/example/uu/fragment_ranking.java
@@ -1,20 +1,0 @@
-package com.example.uu;
-
-import android.os.Bundle;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-
-import androidx.annotation.NonNull;
-import androidx.annotation.Nullable;
-import androidx.fragment.app.Fragment;
-
-public class fragment_ranking extends Fragment {
-    @Nullable
-    @Override
-    public View onCreateView(@NonNull LayoutInflater inflater, @Nullable ViewGroup container, @Nullable Bundle savedInstanceState) {
-        ViewGroup rootview=(ViewGroup) inflater.inflate(R.layout.fragment_ranking,container,false);
-
-        return rootview;
-    }
-}
